requirements.txt
# Telegram Bot
python-telegram-bot==20.7
aiogram==3.3.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Web Framework
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
jinja2==3.1.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Sheets Integration
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utils
python-dotenv==1.0.0
apscheduler==3.10.4
aiofiles==23.2.1
httpx==0.25.2

# Charts and Visualization
plotly==5.18.0
pandas==2.2.0 